Remove DataTable and renaming
