+++ src/components/SplashScreen.vue
