+++ src/main.ts
@@ -2,13 +2,13 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import electron from 'electron';
 import Vue from 'vue';
-import Main from './components/MainPage';
+import App from './components/App';
 import FormField from './components/shared/FormField';
+import SplashScreen from './components/SplashScreen';
 import { initializeReporter } from './modules/reporting';
 import LoggerPlugin from './plugins/logger';
 import vuetify from './plugins/vuetify';
 import router from './router';
-import Splash from './Splash';
 import store from './store';
 
 initializeReporter();
@@ -44,6 +44,6 @@ new Vue({
     });
   },
   render(h) {
-    return this.loaded ? h(Main) : h(Splash);
+    return this.loaded ? h(App) : h(SplashScreen);
   },
 }).$mount('#app');
