+++ src/router/index.js
@@ -1,14 +1,14 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
-import MainPage from '../components/MainPage';
+import App from '../components/App';
 
 Vue.use(VueRouter);
 
 const routes = [
   {
     path: '/',
-    name: 'MainPage',
-    component: MainPage,
+    name: 'App',
+    component: App,
   },
 ];
